拆分设置页按钮，兼容x86架构以便测试
