+++ app/src/main/AndroidManifest.xml
@@ -65,6 +65,15 @@
             android:name=".ui.activity.SettingActivity"
             android:screenOrientation="sensorLandscape" />
         <activity
+            android:name=".ui.activity.PlaySettingsActivity"
+            android:screenOrientation="sensorLandscape" />
+        <activity
+            android:name=".ui.activity.DanmuSettingsActivity"
+            android:screenOrientation="sensorLandscape" />
+        <activity
+            android:name=".ui.activity.PersonalSettingsActivity"
+            android:screenOrientation="sensorLandscape" />
+        <activity
             android:name=".ui.activity.HistoryActivity"
             android:screenOrientation="sensorLandscape" />
         <activity
