+++ app/src/main/java/com/github/tvbox/osc/ui/dialog/ApiDialog.java
@@ -90,8 +90,40 @@ public class ApiDialog extends BaseDialog {
         findViewById(R.id.apiHistory).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                ArrayList<String> history = Hawk.get(HawkConfig.API_HISTORY, new ArrayList<String>());
+                if (history.isEmpty()) {
+                    Toast.makeText(getContext(), "配置历史为空", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+                String current = Hawk.get(HawkConfig.API_URL, "");
+                int idx = 0;
+                if (history.contains(current))
+                    idx = history.indexOf(current);
+                ApiHistoryDialog dialog = new ApiHistoryDialog(getContext());
+                dialog.setTip("配置历史列表");
+                dialog.setAdapter(new ApiHistoryDialogAdapter.SelectDialogInterface() {
+                    @Override
+                    public void click(String value) {
+                        inputApi.setText(value);
+                        Hawk.put(HawkConfig.API_URL, value);
+                        dialog.dismiss();
+                    }
+
+                    @Override
+                    public void del(String value, ArrayList<String> data) {
+                        Hawk.put(HawkConfig.API_HISTORY, data);
+                    }
+                }, history, idx);
+                dialog.show();
+            }
+        });
+
+        // 历史直播按钮
+        findViewById(R.id.apiHistoryLive).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
                 ArrayList<String> history = Hawk.get(HawkConfig.LIVE_API_HISTORY, new ArrayList<String>());
-                if (history.isEmpty()){
+                if (history.isEmpty()) {
                     Toast.makeText(getContext(), "直播历史为空", Toast.LENGTH_SHORT).show();
                     return;
                 }
